custom_components/naim_media_player/config_flow.py
"""Config flow for Naim Control integration."""

from __future__ import annotations

import logging
from typing import Any

import aiohttp
import voluptuous as vol
from homeassistant import config_entries
from homeassistant.const import CONF_IP_ADDRESS, CONF_NAME
from homeassistant.core import HomeAssistant
from homeassistant.data_entry_flow import FlowResult
from homeassistant.exceptions import HomeAssistantError

from .const import DEFAULT_NAME, DOMAIN

_LOGGER = logging.getLogger(__name__)

STEP_USER_DATA_SCHEMA = vol.Schema(
    {
        vol.Required(CONF_IP_ADDRESS): str,
        vol.Optional(CONF_NAME, default=DEFAULT_NAME): str,
    }
)


async def validate_input(hass: HomeAssistant, data: dict[str, Any]) -> dict[str, Any]:
    """Validate the user input allows us to connect."""

    try:
        async with aiohttp.ClientSession() as session:
            async with session.get(
                f"http://{data[CONF_IP_ADDRESS]}:15081/power", timeout=5
            ) as response:
                if response.status != 200:
                    raise CannotConnect

                # Try to parse the response to validate it's a Naim device
                response_data = await response.json()
                if "system" not in response_data:
                    raise InvalidResponse

    except aiohttp.ClientError:
        raise CannotConnect
    except Exception:
        raise InvalidResponse

    # Return info to be stored in the config entry
    return {"title": data[CONF_NAME]}


class NaimControlFlowHandler(config_entries.ConfigFlow, domain=DOMAIN):
    """Handle a config flow for Naim Control."""

    VERSION = 1

    async def async_step_user(
        self, user_input: dict[str, Any] | None = None
    ) -> FlowResult:
        """Handle the initial step."""
        errors: dict[str, str] = {}

        if user_input is not None:
            try:
                info = await validate_input(self.hass, user_input)
            except CannotConnect:
                errors["base"] = "cannot_connect"
            except InvalidResponse:
                errors["base"] = "invalid_response"
            except Exception:  # pylint: disable=broad-except
                _LOGGER.exception("Unexpected exception")
                errors["base"] = "unknown"
            else:
                return self.async_create_entry(title=info["title"], data=user_input)

        return self.async_show_form(
            step_id="user", data_schema=STEP_USER_DATA_SCHEMA, errors=errors
        )


class CannotConnect(HomeAssistantError):
    """Error to indicate we cannot connect."""


class InvalidResponse(HomeAssistantError):
    """Error to indicate we received an invalid response."""
